# Use real-valued instead of complex tensors in Wan2.1 RoPE