@a-r-r-o-w 2025-07-01
@yiyixuxu Can we merge this? There are some numerical differences when running pipeline end-to-end and the final outputs look a bit different. I think this can be expected because the two RoPE implementations have differences on the order of 1e-5 or lower IIRC. The speedup gain from compile compatibility is considerable, so we should definitely try to land this even though there's some numerical difference