@HuggingFaceDocBuilderDev 2025-07-01
The docs for this PR live [here](https://moon-ci-docs.huggingface.co/docs/diffusers/pr_11649). All of your documentation changes will be reflected on that endpoint. The docs are available until 30 days after the last update.